update feedback PR
