+++ ImageFeed/Models/Profile.swift
@@ -1,0 +1,22 @@
+//
+//  Profile.swift
+//  ImageFeed
+//
+//  Created by Anna on 04.03.2024.
+//
+
+import Foundation
+
+struct Profile {
+    let username: String
+    let name: String
+    let loginName: String
+    let bio: String?
+
+    init(username: String, firstName: String, lastName: String, bio: String?) {
+        self.username = username
+        self.name = "\(firstName) \(lastName)"
+        self.loginName = "@\(username)"
+        self.bio = bio
+    }
+}
