+++ ImageFeed/Models/ProfileImage.swift
@@ -1,0 +1,12 @@
+//
+//  ProfileImage.swift
+//  ImageFeed
+//
+//  Created by Anna on 04.03.2024.
+//
+
+import Foundation
+
+struct ProfileImage: Codable {
+    let small: String?
+}
