+++ ImageFeed/Models/ProfileResult.swift
@@ -1,0 +1,22 @@
+//
+//  ProfileResult.swift
+//  ImageFeed
+//
+//  Created by Anna on 04.03.2024.
+//
+
+import Foundation
+
+struct ProfileResult: Codable {
+    let username: String
+    let firstName: String
+    let lastName: String
+    let bio: String?
+    
+    enum CodingKeys: String, CodingKey {
+        case username = "username"
+        case firstName = "first_name"
+        case lastName = "last_name"
+        case bio = "bio"
+    }
+}
