+++ ImageFeed/ServiceAccessLayer/OAuth2/OAuth2Service.swift
@@ -155,69 +155,3 @@ extension URLSession {
         }
 }
 
-enum NetworkError: Error {
-    case badURL
-    case httpStatusCode(Int)
-    case badWebKitResponse
-    case urlRequestError(Error)
-    case urlSessionError
-    case invalidAccessToken
-}
-
-
-//extension URLSession {
-//    func data(
-//        for request: URLRequest,
-//        completion: @escaping (Result<Data, Error>) -> Void
-//    ) -> URLSessionTask {
-//        let fulfillCompletion: (Result<Data, Error>) -> Void = { result in
-//            DispatchQueue.main.async {
-//                completion(result)
-//            }
-//        }
-//
-//        let task = dataTask(with: request, completionHandler: { data, response, error in
-//            if let data = data,
-//               let response = response,
-//               let statusCode = (response as? HTTPURLResponse)?.statusCode
-//            {
-//                if 200..<300 ~= statusCode {
-//                    fulfillCompletion(.success(data))
-//                } else {
-//                    fulfillCompletion(.failure(NetworkError.httpStatusCode(statusCode)))
-//                }
-//            } else if let error = error {
-//                fulfillCompletion(.failure(NetworkError.urlRequestError(error)))
-//            } else {
-//                fulfillCompletion(.failure(NetworkError.urlSessionError))
-//            }
-//        })
-//        task.resume()
-//        return task
-//    }
-//}
-
-
-
-//let task = object(for: request) { [weak self] result in
-//                DispatchQueue.main.async {
-//                    guard let self = self else { return }
-//                    switch result {
-//                    case .success(let body):
-//                        let authToken = body.accessToken
-//                        self.authToken = authToken
-//                        completion(.success(authToken))
-//                    case .failure(let error):
-//                        completion(.failure(error))
-//                    }
-//                    self.task = nil
-//                    self.lastCode = nil
-//                }
-//            }
-
-//let response = result.flatMap { data -> Result<OAuthTokenResponseBody, Error> in
-//                Result {
-//                    try decoder.decode(OAuthTokenResponseBody.self, from: data)
-//                }
-//            }
-//            completion(response)
