+++ ImageFeed/ServiceAccessLayer/OAuth2/OAuth2TokenStorage.swift
@@ -13,11 +13,9 @@ final class OAuth2TokenStorage {
     
     var token: String? {
         get {
-            //return UserDefaults.standard.string(forKey: tokenKey)
             return KeychainWrapper.standard.string(forKey: tokenKey)
         }
         set {
-            //UserDefaults.standard.set(newValue, forKey: tokenKey)
             if let newValue = newValue {
                 KeychainWrapper.standard.set(newValue, forKey: tokenKey)
             } else {
