+++ ImageFeed/ServiceAccessLayer/ProfileService.swift
@@ -8,34 +8,7 @@
 import Foundation
 
 final class ProfileService {
-    struct ProfileResult: Codable {
-        let username: String
-        let firstName: String
-        let lastName: String
-        let bio: String?
-        
-        enum CodingKeys: String, CodingKey {
-            case username = "username"
-            case firstName = "first_name"
-            case lastName = "last_name"
-            case bio = "bio"
-        }
-    }
-    
-    struct Profile {
-        let username: String
-        let name: String
-        let loginName: String
-        let bio: String?
 
-        init(username: String, firstName: String, lastName: String, bio: String?) {
-            self.username = username
-            self.name = "\(firstName) \(lastName)"
-            self.loginName = "@\(username)"
-            self.bio = bio
-        }
-    }
-    
     static let shared = ProfileService()
     var profile: Profile?
     
@@ -85,47 +58,3 @@ final class ProfileService {
         return request
     }
 }
-
-
-
-//        task = urlSession.dataTask(with: request) { [weak self] data, response, error in
-//            guard let self = self else { return }
-//
-//            if let error = error {
-//                completion(.failure(error))
-//                return
-//            }
-//
-//            guard let httpResponse = response as? HTTPURLResponse else {
-//                completion(.failure(NetworkError.badWebKitResponse))
-//                return
-//            }
-//
-//            if httpResponse.statusCode == 401 {
-//                completion(.failure(NetworkError.invalidAccessToken))
-//                return
-//            }
-//
-//            guard 200..<300 ~= httpResponse.statusCode else {
-//                completion(.failure(NetworkError.httpStatusCode(httpResponse.statusCode)))
-//                return
-//            }
-//
-//            do {
-//                guard let data = data else {
-//                    completion(.failure(NetworkError.urlSessionError))
-//                    return
-//                }
-//
-//                let profileResult = try JSONDecoder().decode(ProfileResult.self, from: data)
-//                let profile = Profile(
-//                    username: profileResult.username,
-//                    firstName: profileResult.firstName,
-//                    lastName: profileResult.lastName,
-//                    bio: profileResult.bio
-//                )
-//                completion(.success(profile))
-//            } catch {
-//                completion(.failure(error))
-//            }
-//        }
